[vfs] E: Split Mount Domain Types
Move Mount and Vfs into dedicated modules and nest the private
path-resolution cache under the mount domain that owns it. Preserve the
existing vfs::mount re-exports and all mount-table behavior.
